#416 Refactor constraint about edge end in sequence diagram
The case where trying to create a constructor edge was implemented as a
special case in constraints, where the constraint was only added to the
constraint set if the call could be a legal constructor call. This
divergence from the design, where all constraints are stored in a set,
was creating an obstacle to the future improvement of the constraint
management. With this change, all constraints are implemented as
predicates.